requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.0.3; python_version >= '3.6'
colorama==0.4.4; platform_system == 'Windows'
flask-sqlalchemy==2.5.1
flask==2.0.2
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.2; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
mysql.connector==2.2.9
mysqlclient==2.0.3
python-dotenv==0.19.1
sqlalchemy==1.4.25; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.0.2; python_version >= '3.6'
